+++ bin/check-rabbitmq-queue-drain-time.rb
@@ -43,6 +43,11 @@ class CheckRabbitMQQueueDrainTime < Sensu::Plugin::Check::CLI
          proc: proc(&:to_i),
          default: 15_672
 
+  option :vhost,
+         description: 'Regular expression for filtering the RabbitMQ vhost',
+         short: '-v',
+         long: '--vhost VHOST'
+
   option :user,
          description: 'RabbitMQ management API user',
          long: '--user USER',
@@ -87,6 +92,11 @@ class CheckRabbitMQQueueDrainTime < Sensu::Plugin::Check::CLI
     rescue
       warning 'could not get rabbitmq queue info'
     end
+
+    if config[:vhost]
+      return rabbitmq_info.queues.select { |x| x['vhost'].match(config[:vhost]) }
+    end
+
     rabbitmq_info.queues
   end